% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uiUtil.R
\name{annotate_html}
\alias{annotate_html}
\title{Function that wraps 'match' and 'nomatch' spans around the parts of the string input that do or do not match the pattern input.}
\usage{
annotate_html(string, pattern)
}
\arguments{
\item{string}{the string to annotate}

\item{pattern}{the regular expression to test}
}
\value{
a string with span tags around matching/non-matching sequences
}
\description{
Function that wraps 'match' and 'nomatch' spans around the parts of the string input that do or do not match the pattern input.
}
