% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataBase.R
\name{applyRestriction}
\alias{applyRestriction}
\title{Apply a restriction to a corpus}
\usage{
applyRestriction(corpus, restriction, querystring = NULL)
}
\arguments{
\item{corpus}{The corpus}

\item{restriction}{The restriction}

\item{querystring}{An optional querystring - if set, only matching entries will be returned}
}
\value{
the restricted corpus
}
\description{
Apply a restriction to a corpus
}
