% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corpusBase.R
\name{Base.Count}
\alias{Base.Count}
\title{Count corpus matches}
\usage{
Base.Count(text, querystring)
}
\arguments{
\item{text}{vector of char}

\item{querystring}{the querystring}
}
\value{
Data frame with columns: Token, Frequency, Pairs
}
\description{
Count corpus matches
}
