% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corpusBase.R
\name{Base.Keywords}
\alias{Base.Keywords}
\title{Calculate Keywords}
\usage{
Base.Keywords(wlist.target, wlist.reference, absolute = TRUE)
}
\arguments{
\item{wlist.target}{Wordlist for target corpus}

\item{wlist.reference}{Wordlist for reference corpus}

\item{absolute}{Should words that are rarer in target have negative values?}
}
\value{
Data frame with columns word, target, reference, normalized versions
of target and reference, and keyness
}
\description{
Calculate Keywords
}
