% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corpusBase.R
\name{Base.KWIC}
\alias{Base.KWIC}
\title{Do a KWIC search}
\usage{
Base.KWIC(text, querystring, meta, contextlength = 30)
}
\arguments{
\item{text}{vector of char}

\item{querystring}{the querystring}

\item{meta}{the metadata}

\item{contextlength}{numer of characters for left/right context}
}
\value{
Data frame with colums: all metadata, left context, match, right context
}
\description{
Do a KWIC search
}
