% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corpusBase.R
\name{Base.Wordlist}
\alias{Base.Wordlist}
\title{Create a Wordlist}
\usage{
Base.Wordlist(text, querystring)
}
\arguments{
\item{text}{a vector of chars}

\item{querystring}{the querystring}
}
\value{
Data frame with columns: Token, Frequency, Pairs
}
\description{
Create a Wordlist
}
