% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corpusSetup.R
\name{basicCorpus}
\alias{basicCorpus}
\title{Basic corpus intialization}
\usage{
basicCorpus(corpusdf, KWICcolselect = "ID", type = "line")
}
\arguments{
\item{KWICcolselect}{}
}
\description{
Basic corpus intialization
}
