% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataBase.R
\name{buildQuerystring}
\alias{buildQuerystring}
\title{Returns a querystring}
\usage{
buildQuerystring(searchterm, searchmode, casesensitive)
}
\arguments{
\item{searchterm}{what to search for}

\item{searchmode}{string, word, or regex}

\item{casesensitive}{whether the search should be casesensitive}
}
\value{
the querystring
}
\description{
Returns a querystring
}
\examples{
buildQuerystring("actually", "word", FALSE)
}
