% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadcorpus.R
\name{cacheCorpus}
\alias{cacheCorpus}
\title{Title}
\usage{
cacheCorpus(config, corpusdir = "corpus", metafile = "meta.csv",
  corpusloader = loadCorpus, filelist = NULL, ...)
}
\arguments{
\item{metafile}{}
}
\description{
Title
}
