% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corpusUtil.R
\name{chisq_value}
\alias{chisq_value}
\title{Compute Chi squared Keyness}
\usage{
chisq_value(a, b, ta, tb)
}
\arguments{
\item{a}{Counts in target corpus}

\item{b}{Counts in reference corpus}

\item{ta}{Number of words in target corpus}

\item{tb}{Number of words in reference corpus}
}
\value{
Chi squared keyness
}
\description{
Compute Chi squared Keyness
}
