% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corpusCompareModule.R
\name{compareModule}
\alias{compareModule}
\title{Title}
\usage{
compareModule(input, output, session, config, corpusSelect, restrictions,
  mainCorpus, appControl)
}
\arguments{
\item{appControl}{}
}
\description{
Title
}
