% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corpusUIModule.R
\name{corpusUIHorizontalInput}
\alias{corpusUIHorizontalInput}
\title{Vertical organization for corpus selection and query UI}
\usage{
corpusUIHorizontalInput(id, config, corpusSelect, extraButton = NULL)
}
\arguments{
\item{id}{id}

\item{config}{ShinyConc config}

\item{corpusSelect}{corpus}

\item{extraButton}{whether an extra display button should be displayed}
}
\description{
Vertical organization for corpus selection and query UI
}
