% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corpusUIModule.R
\name{corpusUIInput}
\alias{corpusUIInput}
\title{CorpusUI Module}
\usage{
corpusUIInput(input, output, session, config, corpusSelect,
  restriction = NULL)
}
\arguments{
\item{input}{input}

\item{output}{output}

\item{session}{session}

\item{config}{ShinyConc config}

\item{corpusSelect}{corpus}

\item{restriction}{restriction module (result)}
}
\value{
CorpusUI Module
}
\description{
CorpusUI Module
}
