% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corpusUIModule.R
\name{corpusUIVerticalInput}
\alias{corpusUIVerticalInput}
\title{Vertical organization for corpus selection and query UI}
\usage{
corpusUIVerticalInput(id, config, corpusSelect, editRestrictions = FALSE)
}
\arguments{
\item{id}{id}

\item{config}{ShinyConc config}

\item{corpusSelect}{corpus}

\item{editRestrictions}{whether the restriction handling interface be
displayed}
}
\description{
Vertical organization for corpus selection and query UI
}
