% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uiUtil.R
\name{createClassesMap}
\alias{createClassesMap}
\title{Helper function for preparing datateble output columnDefs}
\usage{
createClassesMap(data, map)
}
\arguments{
\item{data}{the data to be desplayed in the table}

\item{map}{a list with desired output classes as names and the corresponding column names as values}
}
\value{
a map of classes
}
\description{
Helper function for preparing datateble output columnDefs
}
