% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corpusTypeLine.R
\name{filterCorpus.lineCorpus}
\alias{filterCorpus.lineCorpus}
\title{filter a lineCorpus}
\usage{
\method{filterCorpus}{lineCorpus}(corpus, querystring, controls, ...)
}
\arguments{
\item{...}{additional parameters}
}
\value{
filtered corpus
}
\description{
filter a lineCorpus
}
