% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corpusTypePair.R
\name{filterCorpus.pairCorpus}
\alias{filterCorpus.pairCorpus}
\title{filter a pairCorpus}
\usage{
\method{filterCorpus}{pairCorpus}(corpus, querystring, controls, ...)
}
\arguments{
\item{...}{additional parameters}
}
\value{
filtered corpus
}
\description{
filter a pairCorpus
}
