% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corpusType.R
\name{filterCorpus}
\alias{filterCorpus}
\title{Generic method for filtering a corpus object}
\usage{
filterCorpus(corpus, querystring, controls, ...)
}
\arguments{
\item{corpus}{The corpus to filter}

\item{querystring}{The querystring to use for filtering}

\item{...}{additional parameters for implementation}
}
\value{
a filtered corpus
}
\description{
Generic method for filtering a corpus object
}
