% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corpusTypeLine.R
\name{getCounts.lineCorpus}
\alias{getCounts.lineCorpus}
\title{Count occurrences for lineCorpus}
\usage{
\method{getCounts}{lineCorpus}(corpus, querystring, controls, ...)
}
\arguments{
\item{...}{additional parameters}
}
\description{
Count occurrences for lineCorpus
}
