% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corpusTypePair.R
\name{getCounts.pairCorpus}
\alias{getCounts.pairCorpus}
\title{Count occurrences for pairCorpus}
\usage{
\method{getCounts}{pairCorpus}(corpus, querystring, controls, ...)
}
\arguments{
\item{...}{additional parameters}
}
\description{
Count occurrences for pairCorpus
}
