% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corpusType.R
\name{getCounts}
\alias{getCounts}
\title{Generic for deriving corpus counts}
\usage{
getCounts(corpus, querystring, controls, ...)
}
\arguments{
\item{corpus}{The corpus to search}

\item{querystring}{The querystring to use for searching}

\item{...}{additional parameters for implementation}
}
\value{
corpus counts; see Base.count
}
\description{
Generic for deriving corpus counts
}
