% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corpusTypeLine.R
\name{getKWIC.lineCorpus}
\alias{getKWIC.lineCorpus}
\title{Create KWIC for lineCorpus}
\usage{
\method{getKWIC}{lineCorpus}(corpus, querystring, controls, ...)
}
\arguments{
\item{...}{additional parameters}
}
\description{
Create KWIC for lineCorpus
}
