% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corpusTypePair.R
\name{getKWIC.pairCorpus}
\alias{getKWIC.pairCorpus}
\title{Create KWIC for pairCorpus}
\usage{
\method{getKWIC}{pairCorpus}(corpus, querystring, controls, ...)
}
\arguments{
\item{...}{additional parameters}
}
\description{
Create KWIC for pairCorpus
}
