% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corpusType.R
\name{getKWIC}
\alias{getKWIC}
\title{Generic for creating a KWIC view}
\usage{
getKWIC(corpus, querystring, controls, ...)
}
\arguments{
\item{corpus}{The corpus to search}

\item{querystring}{The querystring to use for searching}

\item{...}{additional parameters for implementation}
}
\value{
KWIC view; see Base.KWIC
}
\description{
Generic for creating a KWIC view
}
