% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corpusType.R
\name{getWordcount}
\alias{getWordcount}
\title{Determine the number of words in corpus}
\usage{
getWordcount(corpus, controls, ...)
}
\arguments{
\item{corpus}{The corpus to count}
}
\value{
a number indicating the amount of words in the corpus. If the number of
 words per file is not specified in the corpus (as required by the corpus type),
 ShinyConc will default to using the number of matches for the regular
 expression "\\w+"
}
\description{
Determine the number of words in corpus
}
