% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corpusTypeLine.R
\name{getWordlist.lineCorpus}
\alias{getWordlist.lineCorpus}
\title{Create Wordlist for lineCorpus}
\usage{
\method{getWordlist}{lineCorpus}(corpus, querystring, controls, ...)
}
\arguments{
\item{corpus}{corpus}

\item{querystring}{querystring}

\item{controls}{mode}

\item{...}{additional parameters}
}
\description{
Create Wordlist for lineCorpus
}
