% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corpusTypePair.R
\name{getWordlist.pairCorpus}
\alias{getWordlist.pairCorpus}
\title{Create Wordlist for pairCorpus}
\usage{
\method{getWordlist}{pairCorpus}(corpus, querystring, controls, ...)
}
\arguments{
\item{corpus}{corpus}

\item{querystring}{querystring}

\item{controls}{mode}

\item{...}{additional parameters}
}
\description{
Create Wordlist for pairCorpus
}
