% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corpusType.R
\name{getWordlist}
\alias{getWordlist}
\title{Create Wordlists}
\usage{
getWordlist(corpus, querystring, controls, ...)
}
\arguments{
\item{corpus}{The corpus to search}

\item{querystring}{The querystring to use for searching}

\item{...}{additional parameters for implementation}
}
\value{
a Wordlist; see Base.Wordlist
}
\description{
Create Wordlists
}
