% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataUtil.R
\name{interleave.df}
\alias{interleave.df}
\title{Interleave two data frames}
\usage{
interleave.df(a, b)
}
\arguments{
\item{a}{the first data frame}

\item{b}{the second data frame}
}
\value{
a data frame containing a[1,], b[1,], a[2,], b[2,] and so on
}
\description{
Interleave two data frames
}
