% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataUtil.R
\name{interleave}
\alias{interleave}
\title{Interleave two vectors.}
\usage{
interleave(a, b)
}
\arguments{
\item{a}{the first list}

\item{b}{the second list}
}
\value{
a vector containing a[1], b[1], a[2], b[2] and so on
}
\description{
Interleave two vectors.
}
\examples{
interleave(1:10, 1:10)
}
