% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataUtil.R
\name{isCaseInsensitive}
\alias{isCaseInsensitive}
\title{Evaluates whether a querystring is case-insensitive}
\usage{
isCaseInsensitive(querystring)
}
\arguments{
\item{querystring}{the querystring}
}
\value{
a logi indicating whether the querystring is case-insensitive
}
\description{
Evaluates whether a querystring is case-insensitive
}
