% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadcorpus.R
\name{loadCorpus}
\alias{loadCorpus}
\title{Load corpus}
\usage{
loadCorpus(corpusdir, meta, config = NULL, charset = config$corpusCharset,
  source = config$Corpus$Source, type = config$Corpus$Type, ...)
}
\arguments{
\item{config}{}
}
\description{
Load corpus
}
