% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/appUtil.R
\name{loadOrInstall}
\alias{loadOrInstall}
\title{Load packages and install them if needed}
\usage{
loadOrInstall(packages)
}
\arguments{
\item{packages}{vector of package names}
}
\value{
Nothing
}
\description{
Load packages and install them if needed
}
\examples{
loadOrInstall(c("shiny", "shinyBS"))
}
