% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corpusUtil.R
\name{loglik_value}
\alias{loglik_value}
\title{Compute LogLik Keyness}
\usage{
loglik_value(a, b, ta, tb)
}
\arguments{
\item{a}{Counts in target corpus}

\item{b}{Counts in reference corpus}

\item{ta}{Number of words in target corpus}

\item{tb}{Number of words in reference corpus}
}
\value{
LogLik keyness
}
\description{
Compute LogLik Keyness
}
