% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/queryUIModule.R
\name{queryField}
\alias{queryField}
\title{Query field module server function}
\usage{
queryField(input, output, session)
}
\arguments{
\item{input}{input}

\item{output}{output}

\item{session}{session}

\item{updates}{A reactive value that can be used to modify the module state}
}
\value{
List containing reactives: searchterm, mode, casesensitive, submitButton, querystring
}
\description{
Query field module server function
}
