% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/queryUIModule.R
\name{queryFieldInput}
\alias{queryFieldInput}
\title{Query field module UI function}
\usage{
queryFieldInput(id, label = "Search/Filter", isMain = TRUE,
  useSubmitButton = TRUE)
}
\arguments{
\item{id}{Shiny Module ID}

\item{label}{Query label}

\item{isMain}{Indicate whether this is the main query or an auxilliary one (e.g. for reference corpus selection in keyword view)}

\item{useSubmitButton}{Should the search be interactive or wait for the submit button to be clicked?}
}
\value{
query field module UI
}
\description{
Query field module UI function
}
