% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/queryUIModule.R
\name{queryUpdaterValues}
\alias{queryUpdaterValues}
\title{Returns empty updates reactiveValues for use with Query module}
\usage{
queryUpdaterValues()
}
\value{
reactiveValues with fields searchterm, mode, and casesensitive, all
set to NULL
}
\description{
Returns empty updates reactiveValues for use with Query module
}
