% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corpusUtil.R
\name{quotemeta}
\alias{quotemeta}
\title{Escape regular expression syntax in a string. Copied from http://stackoverflow.com/questions/14836754/is-there-an-r-function-to-escape-a-string-for-regex-characters}
\usage{
quotemeta(string)
}
\arguments{
\item{string}{The string to escape}
}
\value{
the original string ready to be used as a regular expression
}
\description{
Escape regular expression syntax in a string. Copied from http://stackoverflow.com/questions/14836754/is-there-an-r-function-to-escape-a-string-for-regex-characters
}
