% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/restrictionsModule.R
\name{restrictionsField}
\alias{restrictionsField}
\title{Restrictions Module server function}
\usage{
restrictionsField(input, output, session, selectionModule, ...)
}
\arguments{
\item{input}{Input}

\item{output}{Output}

\item{session}{Session}

\item{selectionModule}{selection Module function}

\item{...}{other paramters passed to selectionModule}
}
\description{
Restrictions Module server function
}
