% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/restrictionsModule.R
\name{restrictionsFieldInput}
\alias{restrictionsFieldInput}
\title{Restrictions module UI}
\usage{
restrictionsFieldInput(id, selectionUI, ...)
}
\arguments{
\item{id}{module ID}

\item{...}{other inputs passed to selectionUI}
}
\description{
Restrictions module UI
}
