% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/appUtil.R
\name{runWithRegExCatch}
\alias{runWithRegExCatch}
\title{Run a function and catch any regex errors, returning an error messge data frame instead}
\usage{
runWithRegExCatch(fun, ...)
}
\arguments{
\item{fun}{function executing the regex search}

\item{...}{other arguments passed to fun}
}
\value{
the result of fun, or a data frame containing an error message if the execution of fun causes an regex error
}
\description{
Run a function and catch any regex errors, returning an error messge data frame instead
}
\examples{
runWithRegExCatch(grep, ".*(", "should lead to an error")
}
