% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corpusSearchContextModule.R
\name{searchContextModule}
\alias{searchContextModule}
\title{Title}
\usage{
searchContextModule(input, output, session, config, searchTool, mainCorpus)
}
\arguments{
\item{mainCorpus}{}
}
\description{
Title
}
