% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selectionUIModule.R
\name{selectionField}
\alias{selectionField}
\title{selection module}
\usage{
selectionField(input, output, session, config, corpusSelect)
}
\arguments{
\item{input}{input}

\item{output}{output}

\item{session}{session}

\item{config}{ShinyConc config file}

\item{corpus}{corpus}
}
\description{
selection module
}
