% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selectionUIModule.R
\name{selectionFieldInputFromConfig}
\alias{selectionFieldInputFromConfig}
\title{create selection module ui}
\usage{
selectionFieldInputFromConfig(id, config, corpusSelect)
}
\arguments{
\item{id}{id}

\item{config}{ShinyConc config}

\item{corpusSelect}{corpus}
}
\value{
selection module ui, as specified in the ShinyConc config
}
\description{
create selection module ui
}
