% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/appBase.R
\name{selectorUpdater}
\alias{selectorUpdater}
\title{Create an Updating observer,}
\usage{
selectorUpdater(type, id, input, session, updates)
}
\arguments{
\item{type}{type of the widget (character)}

\item{id}{id of widget}

\item{input}{input}

\item{session}{session}

\item{updates}{updates}
}
\description{
Create an Updating observer,
}
